+++ totalRP3/tools/Locale.lua
@@ -1174,24 +1174,38 @@ Would you like to reload the interface now?]],
 ## Add-on communications improvement
 
  Total RP 3 implements the next version of the Mary Sue Protocol. This improved version bring the following improvements:
- - |cffff8000Profiles are now logged on Blizzard's servers|r when sent to other players. This is so Blizzard can see what's the content of people's RP profile in case of report of abuses, like target harassment or doxxing. |rThis means that you should now treat what you put in your profile as if you were witting that content in /say |cffaaaaaa(Goldshire)|r.
+ - |cffff8000Profiles are now logged on Blizzard's servers|r when sent to other players. This enables Blizzard to view the content of RP profiles in cases of abuse, such as harassment or doxxing. This means you should now treat the contents of your profile as you would public chat in /say, for example. |cffaaaaaa(Goldshire)|r.
  - |cffff8000Cross server and cross faction support with Battle.net friends:|r the add-on can now use Battle.net to transfer data between two Battle.net friends, even if they are from a server that is not connected to yours or if they are from the opposite faction.
- - |cffff8000Improved performances:|r thanks to looser limitations and newer compression algorithms, all data transfers should be faster (sometimes up to 8 times faster for big Extended campaigns).
+ - |cffff8000Improved performance:|r thanks to looser limitations and newer compression algorithms, all data transfer should be faster (sometimes up to 8 times faster for big Extended campaigns).
 
 |cffaaaaaaIt was not possible to make this newer protocol backward compatible with older versions (which will not work with patch 8.0 anyway) and cross add-on communications will only work between people using this newer version of the protocol.|r
 
 ## New logos
 {img:Interface\AddOns\totalRP3\resources\trp3logo.blp:128:64}
-Total RP 3 has a new original logo, to replace the modified game logo (which we obviously did not owned), that was commissioned to [EbonFeathers@Tumblr](https://ebonfeathers.tumblr.com/). Using the theme of classic D&D, this logo showcases that role-playing is all about picking the role _you_ want to play.
+Total RP 3 has a new, original logo, to replace the modified game logo (which we obviously did not own). It was commissioned to [EbonFeathers@Tumblr](https://ebonfeathers.tumblr.com/). Using the theme of classic D&D, this logo showcases that role-playing is all about picking the role you want to play.
 
 A new minimap icon also replaces the older one and showcases a classic D&D die.
 {img:Interface\AddOns\totalRP3\resources\trp3minimap.tga:25:25}
 
-## Other modifications
+## Added
 
-- The map features have been temporarily disabled while we are still working on having them fixed for Battle for Azeroth. The entire world map UI has been completely re-implemented by Blizzard and it requires more or less a complete rewrite of how we handle map stuff.
+- Added a new settings category called Advanced. Changing the settings on this page may break your experience of the add-on, so a warning message will be displayed to warn you when you modify something, and a reset button will allow you to reset all advanced settings to their default values. Amongst these new advanced settings you can find the settings for the broadcast channel, NPC talk prefix, disable the option to remember the last language used between session, and more.
+- Resources added to the browsers: 369 musics, 1698 icons and 178 images from the Battle for Azeroth expansion.
+
+## Modified
+
 - You can no longer set your residence marker inside an instanced zone.
 - Fixed several issues related to patch 8.0.1.
+
+## Fixed
+
+- Added support for other add-ons through the Mary Sue Protocol when using `/trp3 open [playerName]` command.
+
+## Removed
+
+- Map features have been temporarily disabled while we keep working on fixing them for Battle for Azeroth. The entire world map UI has been re-implemented by Blizzard and it requires more or less a complete rewrite of our map code.
+- We have disabled the button to show the residence location of players from the profile page while we are re-implementing the map features for patch 8.0.
+- The system to upvote or downvote profiles have been removed. The system was confusing to new players and was incorrectly used by groups of people to downvote targeted people.
 ]],
 	CL_TOOLTIP = "Create a chat link",
 	CO_ADVANCED_SETTINGS = "Advanced settings",